refactor(guided-tour): improved folder structure
