+++ packages/core/admin/admin/src/components/GuidedTour/Modal/components/Content.js
