+++ packages/core/admin/admin/src/components/GuidedTour/Modal/reducer.js
